requirements.txt
Flask
Flask-Cors
torch
torchvision
numpy
Pillow
opencv-python
grad-cam
